Extract one YANI from spells bigidea back into regular list
Decided not to lump spell ideas into one bigidea, so the one
miscellaneous part of the bigidea that dealt with spell levels
shouldn't be in there. After some digging, found its original author
(K2).

Now that ideas are organized by tag, I may decide to do more like this
in the future and split off more independent bits of bigideas.