Remove punctuation in tags
I tried really hard to make it work with the punctuation in the tags. It
required so much brittle hardcoding, both coming and going, in the
scripts and templates and the names of the html files. Maybe if I ever
refactor this into a full stack web application without as fragile of a
backend I'll take another look at punctuation tags, but for now, they're
out of the picture.

This doesn't remove apostrophes in tags. We'll see how those work.