list-tags.sh: support writing out a CSV file to _data/
I tried to get the frequencies of each tag with pure Liquid template
code, and was not able to because it lacks a filter that can count the
number of times a string appears in an array. So instead, this
workaround: this script already generates the frequencies of each tag.
Add an option that writes them out to a data file.

list.tags.sh --write-csv would be a good pre-commit hook.