Tag rename: "grave" to "gravestone"
"grave" was being used for two ambiguous meanings: the actual terrain
type GRAVE, and the specific grave feature which generates a gravestone
and some buried loot in the Dungeons of Doom. Decided the latter case
can be covered by "random gravestone" in text, and renamed the tag to
make its intention clearer.