Add a "Recent Ideas" page containing the most recent batch of ideas
This is something I've wanted for a while and probably the readers of
the site do too: a way to look at only the ideas added in the last
update, without having to load the all.html megapage. Requires manual
adjusting of the html template each time to update the dates and the new
number.