Remove an idea that seems to fully exist in vanilla already
You can hit $ while deciding where to jump to show valid locations, and
picking one that's too far will cancel it without loss of a turn. So the
idea is suggesting nothing new.

I'd tag it as implemented in vanilla if vanilla had implemented it after
this, but it looks like it existed already by the time this idea was
written down.