Recognize Hack'EM as a variant
This now appears on the variants history graph, so it gets that same
color for its category which is also shared by hackemslashem (the
author).

The category name for use on ideas is "HackEM" with no apostrophe
because the apostrophe isn't valid as a CSS tag, which category names
automatically turn into.