Remove RSS link from <head> of pages
With this in <head>, users' browsers could advertise them a RSS feed of
posts, which is not the format of this jekyll site (the person who
brought this to my attention was confused because they thought it was
like the NHQDB with a relatively constant stream of new ideas added; but
this site updates very infrequently in large chunks). There was never a
functioning RSS feed, so remove the link.