Get rid of unused tag pages
Turns out that when a tag stops being used, its tag page can fall
through the cracks and not be deleted if I don't realize I am deleting
or renaming the last use of a tag. Over the years a bunch of these have
accumulated. This commit gets rid of them.