Update some variant author bios
Cover new variants that people have started and contributed to.

An informal practice I'm starting to use here is: variants that have no
YANIs do not get assigned a color. When such a variant is mentioned on
its author's page, it gets a link to the nethackwiki page for that
variant instead.

This means that there's one author (Noisytoot) who does not have an
author color, because there are no YANIs for notnotdNetHack at this
time.